change integer to bigint, add free ram to server model

+++ db/migrate/20170519234411_create_servers.rb
@@ -3,10 +3,10 @@ class CreateServers < ActiveRecord::Migration[5.0]
     create_table :servers do |t|
       t.string :name
       t.string :hostname
-      t.integer :ram_capacity
-      t.integer :current_ram_usage
-      t.integer :cores_available
-      t.integer :current_core_usage
+      t.bigint :ram_capacity
+      t.bigint :current_ram_usage
+      t.bigint :cores_available
+      t.bigint :current_core_usage
 
       t.timestamps
     end
